1) Dao-module: - Fixed Hibernate - Added JPA with using Hibernate (but we must choose one from them or refactor the error of existing 2 beans of spring-EntityManger); - Add configuration files for Hibernate and JPA; - Fixed Dao-Layer; - Fixed User entity (added first and last); 2) Pom.xml files: - Fixed bugs in pom.xml files of dao module.; - Added some dependencies in dao module and keeper-app; 3) App-module: - Created registration forms - Implemented AuthController for creating users

+++ keeper-core-util/src/main/java/com/keeper/util/CollectorResolver.java
@@ -17,14 +17,14 @@ public class CollectorResolver {
     //<editor-fold desc="makeList">
     private static final List<String>       nullableStringList  = null;
     private static final List<Route>        nullableRouteList   = null;
-    private static final List<Integer>      nullableIdList      = null;
+    private static final List<Long>      nullableIdList      = null;
     private static final List<Location>     nullableLocationList = null;
     private static final List<Coordinate>   nullableCoordList   = null;
     private static final List<User>         nullableUserList    = null;
     private static final List<Task>         nullableTaskList    = null;
 
-    public static List<Integer> makeIdList(final Integer value) {
-        return value != null ? new ArrayList<Integer>() {{ add(value); }} : nullableIdList;
+    public static List<Long> makeIdList(final Long value) {
+        return value != null ? new ArrayList<Long>() {{ add(value); }} : nullableIdList;
     }
 
     public static List<String> makeStringList(final String value) {
