+++ keeper-dao/src/main/java/com/keeper/dao/hibernate/LocationDao.java
@@ -4,8 +4,11 @@ import com.keeper.entity.Location;
 
 /**
  * Created by AlexVasil on 22.03.2017.
+ *
+ * Updated by AlexVasil on 26.03.2017.
+ *
  */
-public interface LocationDao extends GenericDao<Location, Integer>  {
+public interface LocationDao extends GenericDao<Location, Long>  {
