+++ keeper-dao/src/main/java/com/keeper/dao/hibernate/TaskDao.java
@@ -4,7 +4,10 @@ import com.keeper.entity.Task;
 
 /**
  * Created by AlexVasil on 22.03.2017.
+ *
+ * Updated by AlexVasil on 26.03.2017.
+ *
  */
-public interface TaskDao extends GenericDao<Task, Integer> {
+public interface TaskDao extends GenericDao<Task, Long> {
 
 }
