+++ keeper-dao/src/main/java/com/keeper/dao/hibernate/UserDao.java
@@ -4,8 +4,11 @@ import com.keeper.entity.User;
 
 /**
  * Created by AlexVasil on 22.03.2017.
+ *
+ * Updated by AlexVasil on 26.03.2017.
+ *
  */
-public interface UserDao extends GenericDao<User, Integer>  {
+public interface UserDao extends GenericDao<User, Long>  {
