+++ keeper-dao/src/main/java/com/keeper/service/ITaskService.java
@@ -8,13 +8,16 @@ import com.keeper.entity.Task;
 
 /**
  * Default Comment
+ *
+ * Updated by AlexVasil on 26.03.2017.
+ *
  */
 public interface ITaskService {
     Task addTask(Task task);
 
-    Task getTask(Integer id);
+    Task getTask(Long id);
 
     Task updateTask(Task task);
 
-    Task removeTask(Integer id);
+    Task removeTask(Long id);
 }
