+++ keeper-dao/src/main/java/com/keeper/service/IUserService.java
@@ -2,6 +2,9 @@ package com.keeper.service;
 
 /*
  * Created by GoodforGod on 23.03.2017.
+ *
+ * Updated by AlexVasil on 26.03.2017.
+ *
  */
 
 import com.keeper.entity.User;
@@ -12,11 +15,11 @@ import com.keeper.entity.User;
 public interface IUserService {
     User addUser(User user);
 
-    User getUser(Integer id);
+    User getUser(Long id);
     User getUser(String email, String phone);
 
     User updateUser(User user);
 
-    User removeUser(Integer id);
+    User removeUser(Long id);
     User removeUser(String email, String phone);
 }
